Fix swiftlint errors
